routes/serializers.py
from rest_framework import serializers
from .models import *
from django.contrib.auth import authenticate

class UserRegistrationSerializer(serializers.ModelSerializer):
    password = serializers.CharField(write_only=True)

    class Meta:
        model = User
        fields = ['email', 'name', 'password']

    def create(self, validated_data):
        return User.objects.create_user(**validated_data)

class UserLoginSerializer(serializers.Serializer):
    email = serializers.EmailField()
    password = serializers.CharField(write_only=True)
    
    def validate(self, data):
        user = authenticate(email=data['email'], password=data['password'])
        if user and user.is_active:
            return user
        raise serializers.ValidationError("Invalid credentials")

class RegistrationSerializer(serializers.ModelSerializer):
    class Meta:
        model = Registrations
        fields = '__all__'  


class ContactFormSerializer(serializers.ModelSerializer):
    class Meta:
        model = ContactForm
        fields = '__all__' 

class ManufacturerDataSerializer(serializers.ModelSerializer):
    class Meta:
        model = ManufacturerData
        fields = '__all__' 