requirements.txt
asgiref==3.9.1
certifi==2025.7.14
charset-normalizer==3.4.2
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
imageio==2.37.0
joblib==1.5.1
lazy_loader==0.4
networkx==3.5
numpy==2.3.1
packaging==25.0
pillow==11.3.0
PyJWT==2.9.0
PyMySQL==1.1.1
requests==2.32.4
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.16.0
sqlparse==0.5.3
threadpoolctl==3.6.0
tifffile==2025.6.11
urllib3==2.5.0
